# Conflicts:
#	WeSplit/HomeScreenUserControl.xaml.cs
#	WeSplit/bin/Debug/WeSplit.exe
#	WeSplit/bin/Debug/WeSplit.pdb
#	WeSplit/obj/Debug/App.g.cs
#	WeSplit/obj/Debug/App.g.i.cs
#	WeSplit/obj/Debug/DetailScreen.g.cs
#	WeSplit/obj/Debug/DetailScreen.g.i.cs
#	WeSplit/obj/Debug/WeSplit.csproj.CoreCompileInputs.cache
#	WeSplit/obj/Debug/WeSplit.csproj.FileListAbsolute.txt
#	WeSplit/obj/Debug/WeSplit.csproj.GenerateResource.cache
#	WeSplit/obj/Debug/WeSplit.csprojAssemblyReference.cache
#	WeSplit/obj/Debug/WeSplit.exe
#	WeSplit/obj/Debug/WeSplit.g.resources
#	WeSplit/obj/Debug/WeSplit.pdb
#	WeSplit/obj/Debug/WeSplit_MarkupCompile.cache
#	WeSplit/obj/Debug/WeSplit_MarkupCompile.lref

+++ WeSplit/ColorPayConverter.cs
@@ -12,7 +12,7 @@ namespace WeSplit
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            double money = (double)value;
+            float money = (int) value;
             if(money < 0)
             {
                 return "#FFFF0000";
